refactor(hub-common): break out content fetching utils into their own files for easier test mocking
affects: @esri/hub-common
