+++ packages/common/src/content/index.ts
@@ -3,7 +3,9 @@ export * from "./get-family";
 export * from "./compose";
 export * from "./contentUtils";
 export * from "./edit";
-export * from "./fetch";
+export * from "./fetchContent";
+export * from "./fetchHubContent";
+export * from "./modelToHubEditableContent";
 export * from "./get-family";
 export * from "./HubContent";
 export * from "./search";
