+++ packages/common/test/content/fetchContent.test.ts
@@ -7,24 +7,11 @@ import {
   IItemEnrichments,
   cloneObject,
   IPolygonProperties,
-  fetchHubContent,
 } from "../../src";
 import * as _enrichmentsModule from "../../src/items/_enrichments";
 import * as _fetchModule from "../../src/content/_fetch";
-import * as scheduleModule from "../../src/content/manageSchedule";
 import * as documentItem from "../mocks/items/document.json";
 import * as multiLayerFeatureServiceItem from "../mocks/items/multi-layer-feature-service.json";
-import {
-  HOSTED_FEATURE_SERVICE_DEFINITION,
-  HOSTED_FEATURE_SERVICE_GUID,
-  HOSTED_FEATURE_SERVICE_ITEM,
-  MAP_SERVICE_DEFINITION,
-  MAP_SERVICE_GUID,
-  MAP_SERVICE_ITEM,
-  PDF_GUID,
-  PDF_ITEM,
-} from "./fixtures";
-import { MOCK_AUTH, MOCK_NOAUTH_HUB_REQOPTS } from "../mocks/mock-auth";
 
 // mock the item enrichments that would be returned for a multi-layer service
 const getMultiLayerItemEnrichments = () => {
@@ -83,7 +70,7 @@ describe("fetchContent", () => {
       itemId,
       layerId: undefined,
       slug,
-      boundary: null,
+      boundary: null as any,
     };
     requestOpts = {
       portal,
@@ -117,9 +104,9 @@ describe("fetchContent", () => {
       expect(getItemSpy).toHaveBeenCalledWith(itemId, options);
       expect(queryFeaturesSpy).not.toHaveBeenCalled();
       // inspect the results
-      expect(result.item).toEqual(documentItem);
+      expect(result.item).toEqual(documentItem as any);
       expect(result.boundary).toEqual({
-        geometry: null,
+        geometry: null as any,
         provenance: undefined,
       });
       expect(result.statistics).toBeUndefined();
@@ -260,7 +247,7 @@ describe("fetchContent", () => {
           ],
         };
         // expected layer
-        itemEnrichments.layers.push({
+        itemEnrichments.layers?.push({
           id: layerId,
           type: "Feature Layer",
           isView: true,
@@ -327,7 +314,7 @@ describe("fetchContent", () => {
         expect(result.item).toEqual(
           multiLayerFeatureServiceItem as unknown as portalModule.IItem
         );
-        expect(result.viewDefinition.definitionExpression).toBe(
+        expect(result.viewDefinition?.definitionExpression).toBe(
           definitionExpression
         );
         expect(result.recordCount).toBe(count);
@@ -490,9 +477,9 @@ describe("fetchContent", () => {
         );
         expect(queryFeaturesSpy).not.toHaveBeenCalled();
         // inspect the results
-        expect(result.item).toEqual(documentItem);
+        expect(result.item).toEqual(documentItem as any);
         expect(result.boundary).toEqual({
-          geometry: null,
+          geometry: null as any,
           provenance: undefined,
         });
         expect(result.statistics).toBeUndefined();
@@ -527,9 +514,9 @@ describe("fetchContent", () => {
         );
         expect(queryFeaturesSpy).not.toHaveBeenCalled();
         // inspect the results
-        expect(result.item).toEqual(documentItem);
+        expect(result.item).toEqual(documentItem as any);
         expect(result.boundary).toEqual({
-          geometry: null,
+          geometry: null as any,
           provenance: undefined,
         });
         expect(result.statistics).toBeUndefined();
@@ -561,9 +548,9 @@ describe("fetchContent", () => {
       expect(getItemSpy).toHaveBeenCalledWith(itemId, options);
       expect(queryFeaturesSpy).not.toHaveBeenCalled();
       // inspect the results
-      expect(result.item).toEqual(documentItem);
+      expect(result.item).toEqual(documentItem as any);
       expect(result.boundary).toEqual({
-        geometry: null,
+        geometry: null as any,
         provenance: undefined,
       });
       expect(result.statistics).toBeUndefined();
@@ -598,9 +585,9 @@ describe("fetchContent", () => {
         expect(getItemSpy).toHaveBeenCalledWith(itemId, options);
         expect(queryFeaturesSpy).not.toHaveBeenCalled();
         // inspect the results
-        expect(result.item).toEqual(documentItem);
+        expect(result.item).toEqual(documentItem as any);
         expect(result.boundary).toEqual({
-          geometry: null,
+          geometry: null as any,
           provenance: undefined,
         });
         expect(result.statistics).toBeUndefined();
@@ -659,228 +646,13 @@ describe("fetchContent", () => {
           expect(queryFeaturesArg.where).toBeUndefined();
           expect(queryFeaturesArg.portal).toBe(requestOpts.portal);
           // inspect the results
-          expect(result.item).toEqual(
-            multiLayerFeatureServiceItem as portalModule.IItem
-          );
+          expect(result.item).toEqual(multiLayerFeatureServiceItem as any);
           // expect(result.boundary).toEqual({ geometry: null, provenance: undefined })
           expect(result.statistics).toBeUndefined();
-          expect(result.layer.id).toBe(layerId);
+          expect(result.layer?.id).toBe(layerId);
           expect(result.recordCount).toEqual(Infinity);
         });
       });
     });
-  });
-});
-
-describe("fetchHubContent", () => {
-  it("gets feature service content", async () => {
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(HOSTED_FEATURE_SERVICE_ITEM)
-    );
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({
-      metadata: null,
-      server: HOSTED_FEATURE_SERVICE_DEFINITION,
-    });
-
-    const chk = await fetchHubContent(HOSTED_FEATURE_SERVICE_GUID, {
-      portal: MOCK_AUTH.portal,
-      authentication: MOCK_AUTH,
-    });
-    expect(chk.id).toBe(HOSTED_FEATURE_SERVICE_GUID);
-    expect(chk.owner).toBe(HOSTED_FEATURE_SERVICE_ITEM.owner);
-    expect(chk.serverExtractCapability).toBeTruthy();
-    expect(chk.serverQueryCapability).toBeTruthy();
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(HOSTED_FEATURE_SERVICE_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(
-      HOSTED_FEATURE_SERVICE_ITEM
-    );
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual([
-      "metadata",
-      "server",
-    ]);
-  });
-
-  it("gets map service content", async () => {
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(MAP_SERVICE_ITEM)
-    );
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({ metadata: null, server: MAP_SERVICE_DEFINITION });
-
-    const chk = await fetchHubContent(MAP_SERVICE_GUID, {
-      portal: MOCK_AUTH.portal,
-      authentication: MOCK_AUTH,
-    });
-    expect(chk.id).toBe(MAP_SERVICE_GUID);
-    expect(chk.owner).toBe(MAP_SERVICE_ITEM.owner);
-    expect(chk.serverExtractCapability).toBeFalsy();
-    expect(chk.serverQueryCapability).toBeTruthy();
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(MAP_SERVICE_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(MAP_SERVICE_ITEM);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual([
-      "metadata",
-      "server",
-    ]);
-  });
-
-  it("gets non-service content", async () => {
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(PDF_ITEM)
-    );
-    const getServiceSpy = spyOn(featureLayerModule, "getService");
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({ metadata: null });
-
-    const chk = await fetchHubContent(PDF_GUID, {
-      authentication: MOCK_AUTH,
-    });
-    expect(chk.id).toBe(PDF_GUID);
-    expect(chk.owner).toBe(PDF_ITEM.owner);
-    expect(chk.serverExtractCapability).toBeFalsy();
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(PDF_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(PDF_ITEM);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual(["metadata"]);
-  });
-
-  it("normalizes the item type", async () => {
-    spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve({
-        type: "Web Mapping Application",
-        typeKeywords: ["hubSite"],
-      })
-    );
-
-    const chk = await fetchHubContent(PDF_GUID, {
-      authentication: MOCK_AUTH,
-    });
-
-    expect(chk.type).toBe("Hub Site Application");
-  });
-
-  it("should get schedule for request with a token", async () => {
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(PDF_ITEM)
-    );
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({ metadata: null });
-
-    const getScheduleSpy = spyOn(scheduleModule, "getSchedule").and.returnValue(
-      Promise.resolve({
-        mode: "manual",
-      })
-    );
-
-    const chk = await fetchHubContent(PDF_GUID, {
-      portal: MOCK_AUTH.portal,
-      authentication: MOCK_AUTH,
-    });
-
-    // test for schedule
-    expect(chk.schedule).toBeDefined();
-    expect(getScheduleSpy.calls.count()).toBe(1);
-
-    expect(chk.id).toBe(PDF_GUID);
-    expect(chk.owner).toBe(PDF_ITEM.owner);
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(PDF_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(PDF_ITEM);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual(["metadata"]);
-  });
-
-  it("handles when an error is thrown", async () => {
-    spyOn(console, "warn").and.callFake((): any => null);
-
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(PDF_ITEM)
-    );
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({ metadata: null });
-
-    const getScheduleSpy = spyOn(scheduleModule, "getSchedule").and.throwError(
-      "Error getting schedule"
-    );
-
-    const chk = await fetchHubContent(PDF_GUID, {
-      portal: MOCK_AUTH.portal,
-      authentication: MOCK_AUTH,
-    });
-
-    // test for schedule
-    expect(chk.schedule).toBeUndefined();
-    expect(getScheduleSpy.calls.count()).toBe(1);
-
-    expect(chk.id).toBe(PDF_GUID);
-    expect(chk.owner).toBe(PDF_ITEM.owner);
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(PDF_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(PDF_ITEM);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual(["metadata"]);
-  });
-
-  it("should not get schedule for request without a token", async () => {
-    const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
-      Promise.resolve(PDF_ITEM)
-    );
-    const fetchItemEnrichmentsSpy = spyOn(
-      _enrichmentsModule,
-      "fetchItemEnrichments"
-    ).and.returnValue({ metadata: null });
-
-    const getScheduleSpy = spyOn(scheduleModule, "getSchedule").and.returnValue(
-      Promise.resolve({
-        mode: "manual",
-      })
-    );
-
-    const chk = await fetchHubContent(PDF_GUID, MOCK_NOAUTH_HUB_REQOPTS);
-
-    // test for schedule
-    expect(chk.schedule).not.toBeDefined();
-    expect(getScheduleSpy.calls.count()).toBe(0);
-
-    expect(chk.id).toBe(PDF_GUID);
-    expect(chk.owner).toBe(PDF_ITEM.owner);
-
-    expect(getItemSpy.calls.count()).toBe(1);
-    expect(getItemSpy.calls.argsFor(0)[0]).toBe(PDF_GUID);
-    // NOTE: the first call to fetchItemEnrichments is done by fetchContent under the hood,
-    // while the second call is done by fetchHubContent. We only care about the second call here
-    expect(fetchItemEnrichmentsSpy.calls.count()).toBe(2);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[0]).toBe(PDF_ITEM);
-    expect(fetchItemEnrichmentsSpy.calls.argsFor(1)[1]).toEqual(["metadata"]);
   });
 });
